requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
qdrant-client[fastembed]==1.11.0
langchain==0.3.0
langchain-ollama==0.2.0
langchain-community==0.3.0
pypdf==5.0.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
sentence-transformers==3.1.1
torch==2.4.1
numpy<2.0.0
